Added StacSpec, for specifying a stack at a particular commit.
If only the stack name is given, then the commit is assumed to be
the HEAD.  If only the commit is given, then the stack name is
assumed to be the name of the default stack.

I'm not sure if I'll actually use this, but it seems to fit with
PackageSpec and DistributionSpec.